+++ Runtime/Models/GameTelemetryModels.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) 2020 - 2023 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
+using System;
 using System.Runtime.Serialization;
 using UnityEngine.Scripting;
 
@@ -12,5 +13,11 @@ namespace AccelByte.Models
         [DataMember] public string EventName;
         [DataMember] public string EventNamespace;
         [DataMember] public object Payload;
+        [DataMember] public DateTime ClientTimestamp;
+
+        public TelemetryBody() 
+        {
+            ClientTimestamp = DateTime.Now;
+        }
     }
 }