+++ Runtime/Server/DedicatedServer.cs
@@ -1,8 +1,9 @@
-// Copyright (c) 2020 - 2022 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2020 - 2023 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
 using AccelByte.Core;
+using AccelByte.Models;
 
 namespace AccelByte.Server
 {
@@ -29,6 +30,16 @@ namespace AccelByte.Server
         {
             Report.GetFunctionLog(GetType().Name);
             coroutineRunner.Run(session.LoginWithClientCredentials(callback));
+        }
+
+        /// <summary>
+        /// Retrieves the JSON Web Key Set (JWKS) asynchronously.
+        /// </summary>
+        /// <param name="callback">The callback to receive the result containing the JWKS.</param>
+        public void GetJwks(ResultCallback<JwkSet> callback)
+        {
+            Report.GetFunctionLog(GetType().Name);
+            coroutineRunner.Run(session.GetJwks(callback));
         }
     }
 }